+++ lib/cucumber/core.rb
@@ -1,4 +1,4 @@
-require 'cucumber/core/events'
+require 'cucumber/core/event_bus'
 require 'cucumber/core/gherkin/parser'
 require 'cucumber/core/gherkin/document'
 require 'cucumber/core/compiler'
@@ -7,7 +7,7 @@ require 'cucumber/core/test/runner'
 module Cucumber
   module Core
 
-    def execute(gherkin_documents, filters = [], event_bus = Events::Bus.new)
+    def execute(gherkin_documents, filters = [], event_bus = EventBus.new)
       yield event_bus if block_given?
       receiver = Test::Runner.new(event_bus)
       compile gherkin_documents, receiver, filters
