+++ spec/cucumber/core_spec.rb
@@ -272,7 +272,7 @@ module Cucumber
             end
           end
 
-          event_bus = Core::Events::Bus.new
+          event_bus = Core::EventBus.new
           report = Core::Report::Summary.new(event_bus)
           execute [gherkin], [Core::Test::Filters::ActivateStepsForSelfTest.new], event_bus
 
@@ -316,7 +316,7 @@ module Cucumber
           end
           logger = []
 
-          event_bus = Core::Events::Bus.new
+          event_bus = Core::EventBus.new
           report = Core::Report::Summary.new(event_bus)
           execute [gherkin], [WithAroundHooks.new(logger)], event_bus
 
@@ -351,7 +351,7 @@ module Cucumber
           end
         end
 
-        event_bus = Core::Events::Bus.new
+        event_bus = Core::EventBus.new
         report = Core::Report::Summary.new(event_bus)
         execute [gherkin], [ Cucumber::Core::Test::TagFilter.new(['@a']) ], event_bus
 
@@ -370,7 +370,7 @@ module Cucumber
           end
         end
 
-        event_bus = Core::Events::Bus.new
+        event_bus = Core::EventBus.new
         report = Core::Report::Summary.new(event_bus)
         execute [gherkin], [ Cucumber::Core::Test::NameFilter.new([/scenario/]) ], event_bus
